+++ component/dust.go
@@ -1,19 +1,18 @@
 package component
 
-
 // DustComponent represents orbital dust particles from glyph transformation
 // Created when gold is completed during quasar phase
 type DustComponent struct {
-	KineticState // PreciseX/Y, VelX/Y, AccelX/Y (Q16.16)
+	KineticState // PreciseX/Y, VelX/Y, AccelX/Y (Q32.32)
 
 	// Grayscale level preserved from original glyph
 	Level GlyphLevel
 
-	// Target orbit radius, randomized per entity (Q16.16)
-	OrbitRadius int32
+	// Target orbit radius, randomized per entity (Q32.32)
+	OrbitRadius int64
 
-	// Chase boost multiplier, decays over time (Q16.16, Scale = 1.0)
-	ChaseBoost int32
+	// Chase boost multiplier, decays over time (Q32.32, Scale = 1.0)
+	ChaseBoost int64
 
 	// Visual
 	Rune rune
