+++ component/spirit.go
@@ -6,17 +6,17 @@ import "github.com/lixenwraith/vi-fighter/terminal"
 // Position presence is at StartX/StartY to avoid target saturation
 // Actual render position is calculated via Lerp from Start to Target
 type SpiritComponent struct {
-	// Starting position in Q16.16 (where the spirit spawned)
-	StartX, StartY int32
+	// Starting position in Q32.32 (where the spirit spawned)
+	StartX, StartY int64
 
-	// Target position in Q16.16 (convergence point)
-	TargetX, TargetY int32
+	// Target position in Q32.32 (convergence point)
+	TargetX, TargetY int64
 
-	// Animation progress in Q16.16: 0 = start, Scale = complete
-	Progress int32
+	// Animation progress in Q32.32: 0 = start, Scale = complete
+	Progress int64
 
-	// Speed increment per tick in Q16.16 (distance-dependent)
-	Speed int32
+	// Speed increment per tick in Q32.32 (distance-dependent)
+	Speed int64
 
 	// Visual properties
 	Rune       rune
